fix: move createContextByConfig from shared to core
