+++ packages/shared/src/index.ts
@@ -14,7 +14,6 @@ export * from './applyDefaultBuilderOptions';
 export * from './constants';
 export * from './createHook';
 export * from './pluginStore';
-export * from './createContext';
 export * from './utils';
 export * from './fs';
 export * from './generateMetaTags';
